requirements.txt
fastapi==0.100.0
uvicorn==0.22.0
langchain==0.0.240
openai==0.27.8
qdrant-client>=1.6.0,<2.0.0
python-multipart==0.0.6
pydantic==1.10.10
tiktoken==0.5.2
PyPDF2==3.0.1
python-dotenv==1.0.1
unstructured==0.6.7
tabulate==0.8.9
pdf2image==1.16.0